+++ src/main/java/godevenner/divexbackend/portfolio/service/PortfolioEntryServiceImpl.java
@@ -13,7 +13,7 @@ public class PortfolioEntryServiceImpl implements PortfolioEntryService {
 
     private final PortfolioEntryRepository portfolioEntryRepository;
 
-    public List<PortfolioEntry> getPortfolioEntries(Integer userId) {
+    public List<PortfolioEntry> getPortfolioEntries(Long userId) {
         return portfolioEntryRepository.findByPortfolioId(userId);
     }
